requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
entrypoints==0.4
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.31.0
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.6.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.57.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.274
langchain-experimental==0.0.11
langchainplus-sdk==0.0.20
langsmith==0.0.27
looker-sdk==23.8.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pathlib==1.0.1
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.24.1
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.48
sqlalchemy-bigquery==1.8.0
sqlparse==0.4.4
streamlit==1.26.0
streamlit-js-eval==0.1.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.21.2
watchdog==3.0.0
yarl==1.9.2
zipp==3.16.2
